requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
pkg-resources==0.0.0
pytz==2017.2
